+++ src/intersects.c
@@ -6,7 +6,7 @@
 /*   By: jsaarine <jsaarine@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/25 16:16:59 by jsaarine          #+#    #+#             */
-/*   Updated: 2022/09/28 20:29:04 by jsaarine         ###   ########.fr       */
+/*   Updated: 2022/09/30 15:16:06 by jsaarine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,9 @@ void	update_hit_record(t_hit_record *hit, double *distance, int i)
 	}
 }
 
-static void	gg_intersections_per_type(t_context *ctx, double distance, size_t i)
+static void	gg_intersections_per_type(t_context *ctx, double distance, int i)
 {
-	ctx->obj = *(t_object *)vec_get(&ctx->scene, i);
+	ctx->obj = *(t_object *)vec_get(&ctx->scene, (size_t)i);
 	if (ctx->obj.type == SPHERE)
 	{
 		if (intersects_sphere(&ctx->ray, &ctx->obj, &distance))
@@ -49,11 +49,11 @@ static void	gg_intersections_per_type(t_context *ctx, double distance, size_t i)
 
 void	intersects(t_context *ctx, double distance)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	ctx->hit = hit_record_new();
-	while (i < ctx->scene.len)
+	while ((size_t)i < ctx->scene.len)
 	{
 		gg_intersections_per_type(ctx, distance, i);
 		i++;
